scale hub settings
