+++ SS14.Launcher/Views/ScaledControl.xaml
@@ -1,0 +1,1 @@
+<c:LayoutTransformControl xmlns:c="https://github.com/avaloniaui" />
